16x16 AVX2 OK

+++ src/avx2/avx2.h
@@ -2,7 +2,7 @@
 
 /** assembly code **/
 extern void avx2_asm_enum_8x32(__m256i *F, uint64_t alpha_shift, void *buf, uint64_t *num, uint64_t idx);
-extern void avx2_asm_enum_16x16(__m256i *F, uint64_t alpha_shift, void *buf, uint64_t *num, uint32_t idx);
+extern void avx2_asm_enum_16x16(__m256i *F, uint64_t alpha_shift, void *buf, uint64_t *num, uint64_t idx);
 
 size_t avx2_enum_8x32(int n, const uint32_t * const F_,
 			    uint32_t * solutions, size_t max_solutions,
